feat: improve test coverage
config package is completely coverage by tests

+++ pkg/fsutil/fsutil.go
@@ -10,9 +10,13 @@ import (
 // CopyFile copies a file from src to dst.
 func CopyFile(src, dst string) error {
 	srcFile, err := os.Open(src)
+	if errors.Is(err, os.ErrNotExist) {
+		return fmt.Errorf("file %s not found", src)
+	}
 	if err != nil {
 		return err
 	}
+
 	defer srcFile.Close()
 
 	dstFile, err := os.Create(dst)
@@ -21,20 +25,9 @@ func CopyFile(src, dst string) error {
 	}
 	defer dstFile.Close()
 
-	_, err = io.Copy(dstFile, srcFile)
-	if err != nil {
+	if _, err := io.Copy(dstFile, srcFile); err != nil {
 		return err
 	}
 
 	return nil
-}
-
-// CopyFileIsExist copies a file from source to destination only in case if
-// destination file exists.
-func CopyFileIfExist(src, dst string) error {
-	if _, err := os.Stat(src); errors.Is(err, os.ErrNotExist) {
-		return fmt.Errorf("file %s not found", src)
-	}
-
-	return CopyFile(src, dst)
 }
